backend/requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
black==23.1.0
click==8.1.3
colorama==0.4.6
Django==4.1.6
django-cors-headers==3.13.0
djangorestframework==3.14.0
mypy-extensions==0.4.3
packaging==23.0
pathspec==0.11.0
platformdirs==2.6.2
psycopg2-binary~=2.9.3
python-decouple==3.7
pytz==2022.7.1
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
tzdata==2022.7
